# app/Http/Controllers/Api/PageController.php
<?php

namespace App\Http\Controllers\Api;

use App\Http\Controllers\Controller;
use App\Http\Requests\PageRequest;
use App\Http\Resources\PageResource;
use App\Models\Page;
use Illuminate\Http\JsonResponse;
use Illuminate\Http\Request;
use Illuminate\Support\Facades\Storage;
use Illuminate\Support\Str;
use OpenApi\Attributes as OA;

#[OA\Tag(name: 'Pages', description: 'Page management')]
class PageController extends Controller
{
    #[OA\Get(
        path: '/api/pages',
        summary: 'List pages (paginated, searchable, filterable)',
        security: [['sanctum' => []]],
        tags: ['Pages'],
        parameters: [
            new OA\Parameter(name: 'search',   in: 'query', schema: new OA\Schema(type: 'string')),
            new OA\Parameter(name: 'menu_id',  in: 'query', schema: new OA\Schema(type: 'integer')),
            new OA\Parameter(name: 'status',   in: 'query', schema: new OA\Schema(type: 'string', enum: ['draft', 'published'])),
            new OA\Parameter(name: 'trashed',  in: 'query', schema: new OA\Schema(type: 'boolean')),
            new OA\Parameter(name: 'per_page', in: 'query', schema: new OA\Schema(type: 'integer')),
        ],
        responses: [new OA\Response(response: 200, description: 'Paginated pages list')]
    )]
    public function index(Request $request): JsonResponse
    {
        $query = Page::with(['menu', 'creator', 'updater'])
            ->when($request->trashed, fn($q) => $q->onlyTrashed())
            ->when($request->search,  fn($q, $s) => $q->where('title', 'like', "%{$s}%"))
            ->when($request->menu_id, fn($q, $m) => $q->where('menu_id', $m))
            ->when($request->status,  fn($q, $s) => $q->where('status', $s))
            ->latest();

        return api_response(PageResource::collection($query->paginate($request->integer('per_page', 15)))->response()->getData(true));
    }

    #[OA\Post(
        path: '/api/pages',
        summary: 'Create a page',
        security: [['sanctum' => []]],
        tags: ['Pages'],
        requestBody: new OA\RequestBody(
            required: true,
            content: new OA\MediaType(
                mediaType: 'multipart/form-data',
                schema: new OA\Schema(
                    required: ['title', 'body', 'status'],
                    properties: [
                        new OA\Property(property: 'title',        type: 'string'),
                        new OA\Property(property: 'body',         type: 'string'),
                        new OA\Property(property: 'status',       type: 'string', enum: ['draft', 'published']),
                        new OA\Property(property: 'publish_date', type: 'string', format: 'date-time'),
                        new OA\Property(property: 'menu_id',      type: 'integer'),
                        new OA\Property(property: 'cover_image',  type: 'string', format: 'binary'),
                    ]
                )
            )
        ),
        responses: [new OA\Response(response: 201, description: 'Page created')]
    )]
    public function store(PageRequest $request): JsonResponse
    {
        $data = $request->validated();
        $data['created_by'] = $request->user()->id;
        $data['slug'] = $data['slug'] ?? Str::slug($data['title']);

        if ($request->hasFile('cover_image')) {
            $data['cover_image'] = $request->file('cover_image')->store('covers', 'public');
        }

        $page = Page::create($data);
        return api_response(new PageResource($page->load(['menu', 'creator'])), 'Page created', 201);
    }

    #[OA\Get(
        path: '/api/pages/{id}',
        summary: 'Get a page',
        security: [['sanctum' => []]],
        tags: ['Pages'],
        parameters: [new OA\Parameter(name: 'id', in: 'path', required: true, schema: new OA\Schema(type: 'integer'))],
        responses: [new OA\Response(response: 200, description: 'Page detail')]
    )]
    public function show(Page $page): JsonResponse
    {
        return api_response(new PageResource($page->load(['menu', 'creator', 'updater'])));
    }

    #[OA\Post(
        path: '/api/pages/{id}',
        summary: 'Update a page (POST with _method=PUT for file upload)',
        security: [['sanctum' => []]],
        tags: ['Pages'],
        parameters: [new OA\Parameter(name: 'id', in: 'path', required: true, schema: new OA\Schema(type: 'integer'))],
        requestBody: new OA\RequestBody(
            content: new OA\MediaType(
                mediaType: 'multipart/form-data',
                schema: new OA\Schema(properties: [
                    new OA\Property(property: '_method',      type: 'string', example: 'PUT'),
                    new OA\Property(property: 'title',        type: 'string'),
                    new OA\Property(property: 'body',         type: 'string'),
                    new OA\Property(property: 'status',       type: 'string', enum: ['draft', 'published']),
                    new OA\Property(property: 'cover_image',  type: 'string', format: 'binary'),
                ])
            )
        ),
        responses: [new OA\Response(response: 200, description: 'Page updated')]
    )]
    public function update(PageRequest $request, Page $page): JsonResponse
    {
        $data = $request->validated();
        $data['updated_by'] = $request->user()->id;

        if ($request->hasFile('cover_image')) {
            if ($page->cover_image) Storage::disk('public')->delete($page->cover_image);
            $data['cover_image'] = $request->file('cover_image')->store('covers', 'public');
        }

        $page->update($data);
        return api_response(new PageResource($page->load(['menu', 'creator', 'updater'])), 'Page updated');
    }

    #[OA\Delete(
        path: '/api/pages/{id}',
        summary: 'Soft-delete a page',
        security: [['sanctum' => []]],
        tags: ['Pages'],
        parameters: [new OA\Parameter(name: 'id', in: 'path', required: true, schema: new OA\Schema(type: 'integer'))],
        responses: [new OA\Response(response: 200, description: 'Page deleted')]
    )]
    public function destroy(Page $page): JsonResponse
    {
        $page->delete();
        return api_response(null, 'Page deleted');
    }

    #[OA\Post(
        path: '/api/pages/{id}/restore',
        summary: 'Restore a soft-deleted page',
        security: [['sanctum' => []]],
        tags: ['Pages'],
        parameters: [new OA\Parameter(name: 'id', in: 'path', required: true, schema: new OA\Schema(type: 'integer'))],
        responses: [new OA\Response(response: 200, description: 'Page restored')]
    )]
    public function restore(int $id): JsonResponse
    {
        $page = Page::onlyTrashed()->findOrFail($id);
        $page->restore();
        return api_response(new PageResource($page), 'Page restored');
    }
}